+++ NEWS.md
@@ -2,6 +2,7 @@
 - added function for Halley bands
 - added function for population simulation
 - added helper functions for the random generating and applying of age categories
+- added reprasentativity check via Total fertility rate
 
 # mortAAR 1.1.6
 - fixed the error in plotting that occurred after the fix of "aes_string" of ggplot2